refactor: extract specifiers to a package
